Split cytoscape tutorial into two
This is to allow easy expansion of new tutorials, and easier
referencing between modules.

The original tutorial in now called network_construction, though
remains the first and basic introduction. This could lead on to
a new pathway analysis module later.

The second tutorial starts using the GO analysis steps of the
original file, with a cluster analysis step using ClusterMaker2